refactor: move dce tail dispatcher
